% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Compute the marginal likelihood of an observed t-value given a
hypothesis}
\usage{
marginal_likelihood(observed_t, prior = function(x) dcauchy(x, scale =
  sqrt(2)/2), n1, n2)
}
\arguments{
\item{observed_t}{The observed t-value.}

\item{prior}{A function object describing the a-priori plausibility
of all effect sizes. The default is a Cauchy prior as in package
BayesFactor (Morey & Rouder, 2015). Can also be a scalar, in
which case it is assumed that the alternative is a point
hypothesis on Cohen's d (with value `prior`).}

\item{n1}{The sample size in group 1}

\item{n2}{The sample size in group 2}
}
\value{
The marginal likelihood of the observed t-value given the
    prior on Cohen's d.
}
\description{
This function can be used to compute a Bayes factor as the ratio
of two marginal likelihoods.
}
\examples{
## Standard Cauchy prior:
marginal_likelihood(3, n1 = 30, n2 = 30)
## Normal prior - N(0, 0.3)
marginal_likelihood(3, function(x) dnorm(x, 0, 0.3), n1 = 30, n2 = 30)

}
