% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{ncp_from_d}
\alias{ncp_from_d}
\title{Compute the non-centrality parameter for non-central t distribution
from Cohen's d}
\usage{
ncp_from_d(d, n1, n2)
}
\arguments{
\item{d}{Cohen's d}

\item{n1}{Sample size in group 1}

\item{n2}{Sample size in group 2}
}
\value{
The non-centrality parameter
}
\description{
Compute the non-centrality parameter for non-central t distribution
from Cohen's d
}
\details{
See page 7 in Erdfelder, Faul, & Buchner (1996)
}
\examples{
ncp_from_d(0.3, 50, 50)

}
\references{
Erdfelder, E., Faul, F., & Buchner, A. (1996). GPOWER: A general 
power analysis program. Behavior research methods, instruments, & 
computers, 28(1), 1-11.
}
