% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{plot_BF_quantiles}
\alias{plot_BF_quantiles}
\title{Plot the distribution of Bayes factors}
\usage{
plot_BF_quantiles(BF_quantiles, thresholds = NULL, ylim = NULL,
  main = "", axis = c("log", "standard"))
}
\arguments{
\item{BF_quantiles}{A `data.frame` returned by `BF_quantiles`.}

\item{thresholds}{A vector of thresholds to be drawn
horizontally. May for example illustrate some thresholds over
which Bayes factors are deemed to be informative.}

\item{ylim}{A vector of two values indicating the limits of the
y-axis. Given on a log-scale.}

\item{main}{The title of the plot.}

\item{axis}{A character vector indicating what y-axis is to
drawn. Can be "log" for the log-axis, "standard" for
non-logarithmic axis, or both (i.e., c("log", "standard") which
is also the default)}
}
\description{
Plots  quantiles of Bayes factors by sample size.
}
\examples{
bfs <- bf_distribution(effect_size = 0.5, sample_sizes = seq(50, 350, by = 50), n_bayes_factors = 300)
quants <- BF_quantiles(bfs)
plot_BF_quantiles(quants, thresholds = c()
legend("topleft", legend = paste0("PR ", c(80, 50, 20)), pch = 3, col = 3:1, bty = "n")

}
