% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{sample_likelihoods}
\alias{sample_likelihoods}
\title{Generate marginal likelihoods in a given interval of t-values}
\usage{
sample_likelihoods(alternative = function(x) dcauchy(x, scale =
  sqrt(2)/2), n1, n2, from = -6, to = 6)
}
\arguments{
\item{alternative}{A function object. The default is a Cauchy prior
with scaling parameter `sqrt(2) / 2` as is the default in package
`BayesFactor` (Morey & Rouder, 2015). This argument can also be a
scalar number, in which case it is assumed that the alternative
is a point hypothesis on Cohen's d (with Cohen's d = `prior`).}

\item{n1}{The sample size in group 1}

\item{n2}{The sample size in group 2}

\item{from}{The lowest t-value}

\item{to}{The largest t-value}
}
\value{
A data.frame of two columns: Column x contains the t-values, 
  column y contains the marginal likelihoods
}
\description{
Generate marginal likelihoods in a given interval of t-values
}
\examples{
sample_likelihoods(n1 = 30, n2 = 30, from = -3, to = 3)
sample_likelihoods(function(x) dnorm(x, 0, 0.3), n1 = 30, n2 = 30)

}
