% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{visualize_prior}
\alias{visualize_prior}
\title{Visualize a prior on Cohen's d for a Bayesian t-test}
\usage{
visualize_prior(alternative = function(x) dcauchy(x, scale = sqrt(2)/2),
  null = TRUE, from = -4, to = 4, xlab = "Cohen's d",
  frame.plot = FALSE, col = "red", lwd = 3, ...)
}
\arguments{
\item{alternative}{A function object. The default is a Cauchy prior
with scaling parameter `sqrt(2) / 2` as is the default in package
`BayesFactor` (Morey & Rouder, 2015). This argument can also be a
scalar number, in which case it is assumed that the alternative
is a point hypothesis on Cohen's d (with Cohen's d = `prior`).}

\item{null}{Boolean. Plot a line on x = 0 representing the null
hypothesis?}

\item{from}{the left margin of the x-axis}

\item{to}{the right margin of the x-axis}

\item{xlab}{the label of the x-axis}

\item{frame.plot}{Should a frame be drawn?}

\item{col}{The color of the curve of the alternative hypothesis.}

\item{lwd}{The line width}

\item{...}{additional parameters passed to function `curve`}
}
\description{
Visualize a prior on Cohen's d for a Bayesian t-test
}
\examples{
## Standard cauchy prior:
visualize_prior(lty = 2, lwd = 2)
## Wide cauchy prior:
visualize_prior(function(x) dcauchy(x, scale = 1))
## Normal prior with M = 0 and SD = 0.4
visualize_prior(function(x) dnorm(x, 0, 0.4))
## Point prior with Cohen's d = 0.4
visualize_prior(0.4)

}
\references{
Morey, R. D., & Rouder, J. N. (2015). BayesFactor: Computation of
    bayes factors for common designs. Retrieved from
    https://CRAN.R-project.org/package=BayesFactor

Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., & Iverson,
    G. (2009). Bayesian t tests for accepting and rejecting the null
    hypothesis. Psychonomic Bulletin & Review, 16(2), 225–237.
}
