% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{weighted_likelihood}
\alias{weighted_likelihood}
\title{Compute the weighted likelihood of an observed t-value}
\usage{
weighted_likelihood(true_d, observed_t, prior = function(x) dcauchy(x,
  scale = sqrt(2)/2), n1, n2)
}
\arguments{
\item{true_d}{A one-element vector, representing an assumed true 
effect size d, conditioned on which the likelihood is computed.}

\item{observed_t}{The observed t-value.}

\item{prior}{A function object describing the a-priori plausibility
of all effect sizes. The default is a cauchy prior
as in package BayesFactor (Morey & Rouder, 2015)}

\item{n1}{The sample size in group 1}

\item{n2}{The sample size in group 2}
}
\value{
The weighted likelihood of the observed t-value.
}
\description{
For an assumed true effect size (Cohen's d), compute the likelihood
of an observed t-value weighted by the a priori plausbility of this
effect size. This function is primarily used to compute the marginal 
likelihood of the effect size given a continuous hypothesis (e.g., 
a Cauchy prior) by integrating in function `marginal_likelihood`
below.
}
