% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{BF_quantiles}
\alias{BF_quantiles}
\title{Compute quantiles for the distribution of Bayes factors}
\usage{
BF_quantiles(dat, quantiles = c(0.2, 0.5, 0.8))
}
\arguments{
\item{dat}{A `data.frame` returned by `estimate_expected_bf`}
}
\description{
Compute quantiles for the distribution of Bayes factors
}
